Move IndicatorComponentBatch to log
